Update medicamentoUsuario frequency field to float
Changed the frecuencia_horas field from int to float in both the model and schema to allow non-integer values. Updated endpoint to fetch user from token instead of path parameter for listing user medications. Refactored UserController to consistently use user_service attribute.

+++ app/models/medicamentoUsuario.py
@@ -14,7 +14,7 @@ class MedicamentoUsuario(Base):
     idUnidad = Column(Integer, ForeignKey("unidades.id", ondelete="CASCADE"), nullable=False)
     
     dosis = Column(Float, nullable=False)
-    frecuencia_horas = Column(Integer, nullable=False)
+    frecuencia_horas = Column(Float, nullable=False)
     fecha_inicio = Column(DateTime(timezone=True), nullable=False)
     fecha_fin = Column(DateTime(timezone=True), nullable=False)
